functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#### Any version change made here should also be made and tested for the web apps in /app/backend and /app/enrichment

azure-functions == 1.17.0
tiktoken==0.7.0
azure.ai.formrecognizer==3.2.1
azure-storage-blob==12.20.0
azure-core == 1.31.0
lxml == 4.9.2
azure-cosmos == 4.7.0
azure-storage-queue == 12.6.0
nltk == 3.9.1
tenacity == 8.2.3
unstructured[csv,doc,docx,email,html,md,msg,ppt,pptx,text,xlsx,xml] == 0.15.9
pyoo == 1.4
azure-search-documents == 11.4.0b11
beautifulsoup4 == 4.12.2
azure-identity==1.17.1
msgraph-sdk==1.12.0
azure-keyvault-certificates==4.9.0
azure-keyvault-secrets==4.9.0
